requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
azure-identity==1.15.0
azure-cosmos==4.9.0
azure-mgmt-cosmosdb==10.0.0b5
azure-storage-blob==12.19.0
python-dotenv==1.0.1
openai==1.12.0
azure-ai-documentintelligence==1.0.0
aiohttp